+++ packages/terra-date-input/src/terra-dev-site/test/date-input/FilledInDisabledDateInputField.test.jsx
@@ -1,6 +1,5 @@
 import React, { useState } from 'react';
 
-// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import DateInputField from 'terra-date-input/lib/DateInputField';
 
 const Example = () => {