+++ packages/terra-layout/src/terra-dev-site/doc/common/ToolbarExample.jsx
@@ -22,7 +22,7 @@ const content = ({
           className={cx('button')}
           onClick={layoutConfig.toggleMenu}
         >
-Toggle Menu
+          Toggle Menu
         </button>
         )}
     </div>