+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* updated `master` references to `main`.
 
 2.31.0 - (June 2, 2020)
 ------------------