+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -5,6 +5,8 @@ Unreleased
 ----------
 ### Changed
 * Fixed `aria-haspopup` and `link` role  incompatibility issue.
+* updated `master` references to `main`.
+* Minor file changes to comply with eslint rules
 
 2.31.0 - (June 2, 2020)
 ------------------